refactor: move appearance models and view models
The appearance package already owns the domain and the services for this
area, but its models and view models still sat in the shared models,
dialogs and viewmodels packages. Now everything the area needs lives in
one place, and the shared packages keep only what no feature claimed.

The package root stays empty on purpose: SettingsService imports the
appearance domain through it, so anything imported there that reaches
back into the services package would close an import cycle.
