+++ mpvqc/appearance/__init__.py
@@ -1,3 +1,6 @@
 # SPDX-FileCopyrightText: mpvQC developers
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
+
+# Stays empty: SettingsService imports the domain through this package, so anything
+# imported here that reaches back into mpvqc.services closes a cycle.
