+++ mpvqc/appearance/models/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-FileCopyrightText: mpvQC developers
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+from .accent_colors import MpvqcAccentColorModel as MpvqcAccentColorModel
+from .color_scheme_preference import MpvqcColorSchemePreferenceModel as MpvqcColorSchemePreferenceModel
