+++ mpvqc/appearance/viewmodels/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-FileCopyrightText: mpvQC developers
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+from .dialog import MpvqcAppearanceDialogViewModel as MpvqcAppearanceDialogViewModel
+from .palette import MpvqcPaletteViewModel as MpvqcPaletteViewModel
