+++ mpvqc/models/__init__.py
@@ -2,8 +2,6 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from .accent_colors import MpvqcAccentColorModel as MpvqcAccentColorModel
-from .color_scheme_preference import MpvqcColorSchemePreferenceModel as MpvqcColorSchemePreferenceModel
 from .dependencies import MpvqcDependencyModel as MpvqcDependencyModel
 from .export_templates import MpvqcExportTemplateModel as MpvqcExportTemplateModel
 from .import_options import ImportOptionsModel as ImportOptionsModel
