+++ mpvqc/startup.py
@@ -83,6 +83,8 @@ def pin_windows_ui_library() -> None:
 
 
 def import_mpvqc_bindings() -> None:
+    import mpvqc.appearance.models  # ruff: ignore[unused-import]
+    import mpvqc.appearance.viewmodels  # ruff: ignore[unused-import]
     import mpvqc.dialogs  # ruff: ignore[unused-import]
     import mpvqc.enums  # ruff: ignore[unused-import]
     import mpvqc.models  # ruff: ignore[unused-import]
