+++ mpvqc/viewmodels/utility/__init__.py
@@ -4,5 +4,4 @@
 
 from .fonts import MpvqcFontsViewModel as MpvqcFontsViewModel
 from .label_width_calculator import MpvqcLabelWidthCalculatorViewModel as MpvqcLabelWidthCalculatorViewModel
-from .palette import MpvqcPaletteViewModel as MpvqcPaletteViewModel
 from .window import MpvqcWindowViewModel as MpvqcWindowViewModel
