+++ pyproject.toml
@@ -129,16 +129,21 @@ files = [
     "mpvqc/appearance/__init__.py",
     "mpvqc/appearance/domain.py",
     "mpvqc/appearance/injections.py",
+    "mpvqc/appearance/models/__init__.py",
+    "mpvqc/appearance/models/accent_colors.py",
+    "mpvqc/appearance/models/color_scheme_preference.py",
     "mpvqc/appearance/services/__init__.py",
     "mpvqc/appearance/services/color_scheme.py",
     "mpvqc/appearance/services/palette_catalog.py",
+    "mpvqc/appearance/viewmodels/__init__.py",
+    "mpvqc/appearance/viewmodels/dialog.py",
+    "mpvqc/appearance/viewmodels/palette.py",
     "mpvqc/application.py",
     "mpvqc/build.py",
     "mpvqc/close_event_filter.py",
     "mpvqc/datamodels.py",
     "mpvqc/dialogs/__init__.py",
     "mpvqc/dialogs/about.py",
-    "mpvqc/dialogs/appearance.py",
     "mpvqc/dialogs/backup.py",
     "mpvqc/dialogs/comment_types.py",
     "mpvqc/dialogs/edit_input.py",
@@ -167,8 +172,6 @@ files = [
     "mpvqc/jobs.py",
     "mpvqc/logging_utils.py",
     "mpvqc/models/__init__.py",
-    "mpvqc/models/accent_colors.py",
-    "mpvqc/models/color_scheme_preference.py",
     "mpvqc/models/dependencies.py",
     "mpvqc/models/export_templates.py",
     "mpvqc/models/import_options.py",
@@ -275,7 +278,6 @@ files = [
     "mpvqc/viewmodels/utility/__init__.py",
     "mpvqc/viewmodels/utility/fonts.py",
     "mpvqc/viewmodels/utility/label_width_calculator.py",
-    "mpvqc/viewmodels/utility/palette.py",
     "mpvqc/viewmodels/utility/window.py",
     "mpvqc/viewmodels/views/__init__.py",
     "mpvqc/viewmodels/views/footer.py",
